feat(magery): Refactoring: enchantments, spells data

+++ ZuluContent/Skills/AnimalLore.cs
@@ -235,13 +235,13 @@ namespace Server.SkillHandlers
             AddHtml(140, 158, 75, 18, FormatStat(c.ElementalWaterResist), false, false);
 
             AddLabel(50, 176, LabelColor, "Poison Protection");
-            AddHtml(140, 176, 75, 18, FormatStat(c.ElementalPoisonResist), false, false);
+            AddHtml(140, 176, 75, 18, FormatStat((int)c.ElementalPoisonResist), false, false);
 
             AddLabel(50, 194, LabelColor, "Physical Protection");
             AddHtml(140, 194, 75, 18, FormatStat(c.ElementalPhysicalResist), false, false);
 
             AddLabel(50, 212, LabelColor, "Magic Immunity");
-            AddHtml(140, 212, 75, 18, FormatStat(c.PermMagicImmunity), false, false);
+            AddHtml(140, 212, 75, 18, FormatStat((int)c.PermMagicImmunity), false, false);
 
             AddImage(30, 244, 2086);
             AddLabel(50, 242, HeaderColor, "Preferred Foods");
